+++ src/app/pages/Unlock/Unlock.tsx
@@ -6,11 +6,11 @@ import { useDispatch } from 'react-redux';
 import { FormField, IconBase } from 'app/atoms';
 import { StyledButton } from 'app/atoms/StyledButton';
 import { TextButton } from 'app/atoms/TextButton';
+import { useShouldShowV2IntroModal } from 'app/hooks/use-should-show-v2-intro-modal';
 import { ReactComponent as LockFillIcon } from 'app/icons/base/lock_fill.svg';
 import { PlanetsBgPageLayout } from 'app/layouts/planets-bg-page-layout';
 import { getUserTestingGroupNameActions } from 'app/store/ab-testing/actions';
 import { useUserTestingGroupNameSelector } from 'app/store/ab-testing/selectors';
-import { useShouldShowV2IntroModal } from 'app/templates/AppHeader/V2IntroductionModal/hooks/use-should-show-v2-intro-modal';
 import { useFormAnalytics } from 'lib/analytics';
 import { ABTestGroup } from 'lib/apis/temple';
 import { DEFAULT_PASSWORD_INPUT_PLACEHOLDER } from 'lib/constants';
