+++ src/app/pages/Welcome/Welcome.tsx
@@ -7,11 +7,11 @@ import { SocialButton } from 'app/atoms/SocialButton';
 import { StyledButton } from 'app/atoms/StyledButton';
 import { SuspenseContainer } from 'app/atoms/SuspenseContainer';
 import { useABTestingLoading } from 'app/hooks/use-ab-testing-loading';
+import { useShouldShowV2IntroModal } from 'app/hooks/use-should-show-v2-intro-modal';
 import { ReactComponent as ImportedIcon } from 'app/icons/base/imported.svg';
 import { ReactComponent as PlusIcon } from 'app/icons/base/plus.svg';
 import GoogleIconSrc from 'app/icons/google-logo.png';
 import { PlanetsBgPageLayout } from 'app/layouts/planets-bg-page-layout';
-import { useShouldShowV2IntroModal } from 'app/templates/AppHeader/V2IntroductionModal/hooks/use-should-show-v2-intro-modal';
 import { CreatePasswordForm } from 'app/templates/CreatePasswordForm';
 import { GoogleBackupStatusModalContent } from 'app/templates/google-backup-status-modal-content';
 import { ImportSeedForm } from 'app/templates/ImportSeedForm';
